requirements.txt
streamlit
tensorflow
tensorflow_hub
pillow
numpy
